% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checklist.R
\name{get_checklist}
\alias{get_checklist}
\title{Get checklist of native plants around given coordinates}
\usage{
get_checklist(lon = NULL, lat = NULL)
}
\arguments{
\item{lon}{Numeric. Longitude in geographic coordinates}

\item{lat}{Numeric. Latitude in geographic coordinates}
}
\value{
A dataframe
}
\description{
Get a checklist of the native vascular plants around a given point in
the Iberian Peninsula
}
\note{
As the original data (AFLIBER database) have 10-km resolution, the
resulting checklist may include taxa present within 10 km distance of the point or
polygon.
}
\examples{
head(get_checklist(lon = -5, lat = 40))

}
