% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs.R
\docType{data}
\name{locs}
\alias{locs}
\title{Taxa distributions}
\format{
\subsection{\code{locs}}{

A data frame with 1,824,549 rows and 5 columns:
\describe{
\item{Taxon}{taxon}
\item{Family}{family}
\item{GBIF_id}{GBIF ID}
\item{lng}{longitude in decimal degrees}
\item{lat}{latitude in decimal degrees}
}
}
}
\source{
\url{https://doi.org/10.1111/geb.13363}
}
\usage{
locs
}
\description{
Taxa distributions
}
\keyword{datasets}
