% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.around.me-package.R
\docType{package}
\name{plants.around.me-package}
\alias{plants.around.me}
\alias{plants.around.me-package}
\title{plants.around.me: Find Native Plants Around Me}
\description{
Get information on native plants found around a given location. Currently only available for the Iberian Peninsula.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/plants.around.me}
  \item Report bugs at \url{https://github.com/Pakillo/plants.around.me/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
